+++ main.cpp
@@ -31,8 +31,8 @@ LRESULT CALLBACK WndProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam)
 		}
 		case WM_SIZE:
 		{
-			uint32_t width = LOWORD(lParam);
-			uint32_t height = HIWORD(lParam);
+			float width = LOWORD(lParam);
+			float height = HIWORD(lParam);
 			uaudio::imgui::window.SetSize(ImVec2(width, height));
 			break;
 		}
